requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
decorator==4.4.2
distro==1.9.0
frozenlist==1.4.1
gTTS==2.5.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.4.9
moviepy==1.0.3
multidict==6.0.5
numpy==1.26.4
openai==0.28.0
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
pillow==10.3.0
proglog==0.1.10
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
pygame==2.5.2
pyttsx3==2.90
requests==2.31.0
sniffio==1.3.1
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
yarl==1.9.4
